Reflect comments
