+++ cmd/npv/app/helper_k8s.go
@@ -6,11 +6,13 @@ import (
 	"fmt"
 	"math/rand/v2"
 	"strconv"
+	"strings"
 
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 	"k8s.io/apimachinery/pkg/fields"
+	"k8s.io/apimachinery/pkg/types"
 	"k8s.io/client-go/dynamic"
 	"k8s.io/client-go/kubernetes"
 	ctrl "sigs.k8s.io/controller-runtime"
@@ -21,6 +23,14 @@ var (
 	cachedIdentityEndpoints map[int][]*unstructured.Unstructured
 	cachedIdentityExample   = make(map[int]*unstructured.Unstructured)
 )
+
+func parseNamespacedName(nn string) (types.NamespacedName, error) {
+	li := strings.Split(nn, "/")
+	if len(li) != 2 {
+		return types.NamespacedName{}, errors.New("input is not NAMESPACE/NAME")
+	}
+	return types.NamespacedName{Namespace: li[0], Name: li[1]}, nil
+}
 
 func createK8sClients() (*kubernetes.Clientset, *dynamic.DynamicClient, error) {
 	config, err := ctrl.GetConfig()
@@ -73,6 +83,13 @@ func getPodIdentity(ctx context.Context, d *dynamic.DynamicClient, namespace, na
 	}
 
 	return identity, nil
+}
+
+func getSubjectNamespace() string {
+	if rootOptions.allNamespaces {
+		return ""
+	}
+	return rootOptions.namespace
 }
 
 func selectSubjectPods(ctx context.Context, clientset *kubernetes.Clientset, name, selector string) ([]*corev1.Pod, error) {
